working on array_t

+++ hw2/Container_t.h
@@ -5,12 +5,12 @@ public:
 
 	Container_t(): size(0) {}
 
-	int count() const { return size; }
+	size_t count() const { return size; }
 	virtual T* find(const T& element) const;
 	
 	virtual void insert(const T& element) = 0;
-	virtual void append(const T& element, int index) = 0;
-	virtual void prepend(const T& element, int index) = 0;
+	virtual void append(const T& element, size_t index) = 0;
+	virtual void prepend(const T& element, size_t index) = 0;
 
 	virtual T* remove(const T& element) = 0;
 	void removeAndDelete(const T& element) { delete remove(element); }
@@ -19,6 +19,6 @@ public:
 	virtual void removeAndDeleteAll() = 0;
 
 private:
-	int size;
+	size_t size;
 
 };
